Spring_2025_Projects/Fullstack-Multimodal-RAG-System-main/pyproject.toml
[project]
name = "mm-rag-lib"
version = "0.1.0"
description = ""
authors = [
    {name = "Viswanath Raju Indukuri"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = "^3.12"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "selenium (>=4.31.0,<5.0.0)",
    "bs4 (>=0.0.2,<0.0.3)",
    "lxml (>=5.3.2,<6.0.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pydantic (>=2.11.3,<3.0.0)",
    "email-validator (>=2.2.0,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "python-jose[cryptography] (>=3.4.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "google-cloud-storage (>=3.1.0,<4.0.0)",
    "langchain-openai (>=0.3.14,<0.4.0)",
    "langchain (>=0.3.24,<0.4.0)",
    "langgraph (>=0.3.34,<0.4.0)",
    "langchain-community (>=0.3.22,<0.4.0)",
    "streamlit (>=1.44.1,<2.0.0)",
    "langchain-upstage (>=0.6.0,<0.7.0)",
    "langchain-core (>=0.3.56,<0.4.0)",
    "google-generativeai (>=0.8.5,<0.9.0)",
    "langchain-text-splitters (>=0.3.8,<0.4.0)",
    "pymupdf4llm (>=0.0.21,<0.0.22)",
    "langchain-opentutorial (>=0.0.8,<0.0.9)",
    "pinecone[grpc] (>=6.0.2,<7.0.0)",
    "langchain-pinecone (>=0.2.5,<0.3.0) ; python_version >= \"3.12\" and python_version < \"3.14\"",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
